requirements.txt
click==8.1.7
diffusers==0.25.0
numpy==1.24.1
perlin_noise==1.12
Pillow==10.2.0
scikit_learn==1.2.2
transformers==4.36.2
accelerate
